requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
Flask-Migrate==4.0.4
python-dotenv==1.0.0
